guile: SunOS specific changes - don't declare STACKBOTTOM unless USERLIMIT is given.
From Brad Forschinger in PR pkg/49787

Also from both boehm-gc commit, where there is a longer description of issues
in issue #122:
https://github.com/ivmai/bdwgc/issues/122
https://github.com/ivmai/bdwgc/commit/81cae56522e601fc2e9e9bab08c87d1302c118fb

we have the additions here (adding the definition of STACKBOTTOM if USERLIMIT
otherwise use HEURISTIC2), but not the removal of the first part, which
declares STACKBOTTOM unconditionally.

Reportedly fixes build on SunOS 5.10 sun4u.