bootstrap: Back to default $pkgdbdir location to $varbase
Default writable data should not be in direct below prefix,
and the main issue is preventing to take over other package system db dir.

Change default $varbase to "/var", being match with README.
Change to stop bootsrap if $prefix or $pkgdbdir exists.
Change to stop if $pkgdbdir does not match with current installation
with "bootstrap --update".