[zstd] Fix to use variable in pkg-config file
PKGCONFIG_OVERRIDE will duplicate "-L@LIBDIR@" with
"$[COMPILER_RPATH_FLAG}@LIBDIR@", but this pkg's Makefile only try @LIBDIR@
once at a line with sed, so one @LIBDIR@ remain in the result of pc file.

This patch change to use ${libdir} instead of @LIBDIR@,
it is defined above with @LIBDIR@, and it will avoid such issue.