Add two patches so that this at least semi-works when the inet6 option is used:
 * Use correct sockaddr length when doing getnameinfo() for inet6,
   so we avoid an early return with "permanent failure" from getnameinfo()
 * Use temp variables for walking the address lists so that we avoid trying
   freeaddrinfo(NULL) and getting SEGV

This still isn't fully baked and backward compatible: with the
inet6 option turned on, on NetBSD the conserver process only opens
an inet6 server socket and no longer serves an inet socket (a
Linuxism, I suspect), making it troublesome to interoperate with
older versions of conserver or installations on hosts without IPv6
connectivity.

PKGREVISION bumped.