Update to 5.3.0.3
* Use internal harfbuzz and graphite2
  LibreOffice 5.3.0 requires harfbuzz with graphite2 support.
  Enabling graphite2 in fonts/harfbuzz breaks graphics/librsvg, so
  I cannot enable graphite2 in fonts/harfbuzz.

* Explicitly specify static libraries of harfbuzz and graphite2
  with path to fix a linking of libvcllo.so

* Add some comments to patches

Changelog:
Highlights

LibreOffice 5.3 offers a number of interesting new features in
every area: a new cross-platform text layout engine that uses
HarfBuzz for consistent text layout on all platforms, with significant
advantages across languages and alphabets; a revised Help menu,
with new quick links to user guides and community support forums,
for an improved user experience; and better import/export filters
to new and legacy MS Office documents.

Writer now supports Table Styles, for applying formatting to a
table which is preserved when you make edits to it; a new Page Deck
in the sidebar lets the user quickly customise page settings without
having to go through a separate dialog box; and a new Go to Page
Box makes it possible to jump to another page in the document with
just a few keystrokes.

Calc provides a new set of default cell styles, with greater variety
and better names than in previous releases; in fresh installations,
"enable wildcards in formulas" is now the default option, rather
than regular expressions, to improve compatibility with other
spreadsheet software; and a new text entry box lets the user narrow
down the functions he is looking for, and simplifies the search
for the right one.

Impress now opens with a template selector, to get the user off to
a quick start; and a new Slide Properties Deck is now available in
the sidebar while in slide master mode.